requirements.txt
apache-airflow-providers-apache-spark>=4.10.0
beautifulsoup4==4.14.3
Requests==2.32.5
pyspark==4.0.1
pandas>=2.0.0